feat 搜索，up主信息，视频评论列表，关于界面

+++ app/src/main/java/com/a10miaomiao/bilimiao/entity/Page.kt
@@ -1,7 +1,7 @@
 package com.a10miaomiao.bilimiao.entity
 
 data class Page(
-    val cid: Double,
+    val cid: Long,
     val dimension: Dimension,
     val dm: Dm,
     val dmlink: String,
